requirements.txt
pygame==2.5.2
numpy==1.26.3
torch==2.2.0+cu121 --index-url https://download.pytorch.org/whl/cu121
torchrl==0.3.0
scipy==1.12.0
seaborn==0.13.2
gymnasium==0.29.1